chore: reformatted the code style into filters and sort tabs
